Add RegisterMsgServiceDesc
